+++ src/ru/netology/bookkeeping/TaxService.java
@@ -1,8 +1,10 @@
 package ru.netology.bookkeeping;
 
+import java.math.BigDecimal;
+
 public class TaxService {
 
-    public void payOut(double taxAmount) {
+    public void payOut(BigDecimal taxAmount) {
         System.out.printf("Уплачен налог в размере %.2f%n", taxAmount);
     }
 }
